requirements.txt
addict==2.4.0
einops
faiss-gpu
matplotlib
munkres
numba
numpy
pandas
pathlib
Pillow
pyarrow
scikit-learn
scipy
Shapely==1.8.1.post1
six==1.15.0
timm==0.4.12
tqdm
PyYAML
